Fixed URL's, repositories

+++ src/main/java/com/frontwit/app/utils/OrderHelper.java
@@ -7,11 +7,11 @@ import com.frontwit.app.entities.Order;
  * Created by Patryk on 09.04.2016.
  */
 public class OrderHelper {
-//TODO modify this piece of shit
-    public static Order assembleInstanceOfOrder(TZamowieniaEntity zam, int amount, String clientName){
+    //TODO modify this piece of shit
+    public static Order assembleInstanceOfOrder(TZamowieniaEntity zam, int amount, String clientName) {
 
         Order order = new Order();
-        order.setId(zam.getId());
+        order.setId((long) zam.getId());
         order.setName(zam.getNumer());
         order.setCustomerId(1);  //todo tutaj ma byc jakos sprawdzane, czy w bazie danych jest dane nazwisko, jak tak to wyszukaj po id
         order.setDate(zam.getData_z());
